Revert "GRP-732 changed bfType uint8 -> uint32"
This reverts commit 1dfbf5c306a7492310fd29c320f8e7a88559576a.

+++ reels/group.go
@@ -1,7 +1,7 @@
 package reels
 
 type BuyFeatureReelsGroup interface {
-	BuyFeatureReels(bfType uint32) ReelsGenerator
+	BuyFeatureReels(bfType uint8) ReelsGenerator
 	BonusReels() ReelsGenerator
 }
